move types into their own file
